Move compressed HDU code inside a dedicated subpackage
