+++ astropy/io/fits/hdu/compressed/__init__.py
@@ -1,0 +1,3 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+
+from .compressed import *
